+++ ra4-hal/src/lib.rs
@@ -47,31 +47,6 @@ pub(crate) use ra4m1_ctpac as pac;
 
 use crate::{mcu_info::McuInfo, write_protect::ProtectedPeripheral as _};
 
-/// Common peripheral operating modes.
-pub mod mode {
-    trait SealedMode {}
-
-    /// Operating mode for a peripheral.
-    #[allow(private_bounds)]
-    pub trait Mode: SealedMode {}
-
-    macro_rules! impl_mode {
-        ($name:ident) => {
-            impl SealedMode for $name {}
-            impl Mode for $name {}
-        };
-    }
-
-    /// Synchronous mode.
-    pub struct Blocking;
-
-    /// Async mode.
-    pub struct Async;
-
-    impl_mode!(Blocking);
-    impl_mode!(Async);
-}
-
 /// System clock configuration
 ///
 /// # Notes