+++ ra4-hal/src/osm/ofs1.rs
@@ -136,16 +136,15 @@ impl Ofs1 {
     /// #[unsafe(link_section = ".ofs1")]
     /// static OFS1: Ofs1 = Ofs1::default();
     /// ```
-    #[cfg(feature = "hoco_48mhz")]
     pub const fn default() -> Self {
-        Self::new::<LvdasOff, Vdsel190, HocoOn, Hoco48Mhz>()
-    }
-    #[cfg(feature = "hoco_32mhz")]
-    pub const fn default() -> Self {
-        Self::new::<LvdasOff, Vdsel190, HocoOn, Hoco32Mhz>()
-    }
-    #[cfg(feature = "hoco_64mhz")]
-    pub const fn default() -> Self {
-        Self::new::<LvdasOff, Vdsel190, HocoOn, Hoco64Mhz>()
+        cfg_if::cfg_if! {
+            if #[cfg(feature = "hoco_32mhz")] {
+                Self::new::<LvdasOff, Vdsel190, HocoOn, Hoco32Mhz>()
+            } else if #[cfg(feature = "hoco_48mhz")] {
+                Self::new::<LvdasOff, Vdsel190, HocoOn, Hoco48Mhz>()
+            } else if #[cfg(feature = "hoco_64mhz")] {
+                Self::new::<LvdasOff, Vdsel190, HocoOn, Hoco64Mhz>()
+            }
+        }
     }
 }